requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain-openai
langchain-qdrant
langchain-core
qdrant-client
python-dotenv
pymupdf
pandas
beautifulsoup4
pillow
pytesseract
langchain-experimental
xlrd
azure-storage-blob
fastapi
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.1.0
